+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
@@ -5,7 +7,8 @@ import 'package:flutter/rendering.dart';
 import 'package:image_gallery_saver/image_gallery_saver.dart';
 
 import 'package:permission_handler/permission_handler.dart';
-
+import 'package:path_provider/path_provider.dart';
+import 'package:dio/dio.dart';
 import 'dart:ui' as ui;
 
 void main() => runApp(MyApp());
@@ -58,24 +61,57 @@ class _MyHomePageState extends State<MyHomePage> {
                 ),
               ),
               Container(
+                padding: EdgeInsets.only(top: 15),
                 child: RaisedButton(
-                  onPressed: _saved,
-                  child: Text("保存到相册"),
+                  onPressed: _saveScreen,
+                  child: Text("保存页面图片到相册"),
                 ),
-                width: 100,
-                height: 50,
-              )
+                width: 200,
+                height: 44,
+              ),
+              Container(
+                padding: EdgeInsets.only(top: 15),
+                child: RaisedButton(
+                  onPressed: _getHttp,
+                  child: Text("保存网络图片到相册"),
+                ),
+                width: 200,
+                height: 44,
+              ),
+              Container(
+                padding: EdgeInsets.only(top: 15),
+                child: RaisedButton(
+                  onPressed: _saveVideo,
+                  child: Text("保存网络视频到相册"),
+                ),
+                width: 200,
+                height: 44,
+              ),
             ],
           ),
         ));
   }
 
-  _saved() async {
+  _saveScreen() async {
     RenderRepaintBoundary boundary =
         _globalKey.currentContext.findRenderObject();
     ui.Image image = await boundary.toImage();
     ByteData byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-    final result = await ImageGallerySaver.save(byteData.buffer.asUint8List());
+    final result = await ImageGallerySaver.saveImage(byteData.buffer.asUint8List());
     print(result);
   }
+
+  _getHttp() async {
+    var response = await Dio().get("https://ss0.baidu.com/94o3dSag_xI4khGko9WTAnF6hhy/image/h%3D300/sign=a62e824376d98d1069d40a31113eb807/838ba61ea8d3fd1fc9c7b6853a4e251f94ca5f46.jpg", options: Options(responseType: ResponseType.bytes));
+    final result = await ImageGallerySaver.saveImage(Uint8List.fromList(response.data));
+    print(result);
+  }
+
+ _saveVideo() async {
+    var appDocDir = await getTemporaryDirectory();
+    String savePath = appDocDir.path + "/temp.mp4";
+    await Dio().download("http://clips.vorwaerts-gmbh.de/big_buck_bunny.mp4", savePath);
+    final result = await ImageGallerySaver.saveFile(savePath);
+    print(result);
+ }
 }