+++ lib/image_gallery_saver.dart
@@ -9,10 +9,18 @@ class ImageGallerySaver {
 
   /// save image to Gallery
   /// imageBytes can't null
-  static Future save(Uint8List imageBytes) async {
+  static Future saveImage(Uint8List imageBytes) async {
     assert(imageBytes != null);
     final result =
     await _channel.invokeMethod('saveImageToGallery', imageBytes);
+    return result;
+  }
+
+  /// Save the PNG，JPG，JPEG image or video located at [file] to the local device media gallery.
+  static Future saveFile(String file) async {
+    assert(file != null);
+    final result =
+    await _channel.invokeMethod('saveFileToGallery', file);
     return result;
   }